LayoutTests/js/sequence-iterator-protocol-2-expected.txt

CONSOLE MESSAGE: Using an array but with objects with valueOf()
CONSOLE MESSAGE: 1,2,3,1,2,3
CONSOLE MESSAGE: 1,2,3,1,2,3
CONSOLE MESSAGE: callCount: 3
CONSOLE MESSAGE: 1,2,3,1,2,3
CONSOLE MESSAGE: 1,2,3,1,2,3
CONSOLE MESSAGE: callCount: 0
PASS successfullyParsed is true

TEST COMPLETE
